(1) Adding a short delay to catch fast game traversal.
(2) Resetting a connection if the position has changed,
but the server didn't response yet. This is a
significant gain in performance.

(3) Caching the server values. Also this is a big gain
in performance, and this behaviour is user-friendly,
because he will see now the results immediately if
a position is already queried.

Cheers,
Gregor